+++ lib/src/model/episode.dart
@@ -3,17 +3,29 @@
 
 /// A class representing an individual episode.
 class Episode {
+  /// The episode unique identifier.
   final String guid;
+
+  /// The episode title.
   final String title;
+
+  /// The episode description.
   final String description;
+
+  /// The episode URL.
   String link;
+
+  /// Publication date of the episode.
   final DateTime publicationDate;
 
   // Media fields
   String contentUrl;
 
   // iTunes specific fields
+  /// Episode author.
   String author;
+
+  /// Length of the episode as a [Duration].
   final Duration duration;
 
   Episode.of(this.guid, this.title, this.description, this.link,